+++ src/test/java/org/sonar/plugins/delphi/pmd/rules/SwallowedExceptionsRuleTest.java
@@ -1,5 +1,6 @@
 package org.sonar.plugins.delphi.pmd.rules;
 
+import static org.sonar.plugins.delphi.utils.conditions.RuleKey.ruleKey;
 import static org.sonar.plugins.delphi.utils.conditions.RuleKeyAtLine.ruleKeyAtLine;
 
 import org.junit.jupiter.api.Test;
@@ -88,9 +89,7 @@ class SwallowedExceptionsRuleTest extends BasePmdRuleTest {
 
     execute(builder);
 
-    assertIssues()
-        .hasSize(1)
-        .areExactly(1, ruleKeyAtLine("SwallowedExceptionsRule", builder.getOffset() + 9));
+    assertIssues().areExactly(1, ruleKeyAtLine("SwallowedExceptionsRule", builder.getOffset() + 9));
   }
 
   @Test
@@ -112,7 +111,7 @@ class SwallowedExceptionsRuleTest extends BasePmdRuleTest {
 
     execute(builder);
 
-    assertIssues().isEmpty();
+    assertIssues().areNot(ruleKey("SwallowedExceptionsRule"));
   }
 
   @Test
@@ -135,7 +134,7 @@ class SwallowedExceptionsRuleTest extends BasePmdRuleTest {
 
     execute(builder);
 
-    assertIssues().isEmpty();
+    assertIssues().areNot(ruleKey("SwallowedExceptionsRule"));
   }
 
   @Test