+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `NoSonarRule` analysis rule, which flags usages of the `NOSONAR` marker.
+
 ## [0.28.0] - 2021-02-02
 
 ### Added