+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.29.0] - 2021-03-23
+
 ### Added
 
 - `NoSonarRule` analysis rule, which flags usages of the `NOSONAR` marker.
@@ -674,7 +676,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - False-positives in 13 rules.
 - Logic errors in existing "semantic analysis" that caused scan failures.
 
-[Unreleased]: https://github.com/integrated-application-development/sonar-delphi/compare/v0.28.0...HEAD
+[Unreleased]: https://github.com/integrated-application-development/sonar-delphi/compare/v0.29.0...HEAD
+[0.29.0]: https://github.com/integrated-application-development/sonar-delphi/compare/v0.28.0...v0.29.0
 [0.28.0]: https://github.com/integrated-application-development/sonar-delphi/compare/v0.27.0...v0.28.0
 [0.27.0]: https://github.com/integrated-application-development/sonar-delphi/compare/v0.26.0...v0.27.0
 [0.26.0]: https://github.com/integrated-application-development/sonar-delphi/compare/v0.25.0...v0.26.0